restructure code into separate files
