+++ src/index.ts
@@ -1,3 +1,8 @@
+import { decoder, encoder, getCharacterLength, getNextCharWidth, getString, getUtf8Char } from "./endecoder";
+import { InputType, InputTypes } from "./inputTypes";
+import { FnReturingParserIterator, Parser, ParserState, Err, Ok, ResultType } from "./parser";
+import { updateData, updateError, updateParserState, updateResult } from "./update";
+
 // Caching compiled regexs for better performance
 const reDigit = /[0-9]/;
 const reDigits = /^[0-9]+/;
@@ -6,311 +11,7 @@ const reLetters = /^[a-zA-Z]+/;
 const reWhitespaces = /^\s+/;
 const reErrorExpectation = /ParseError.+Expecting/;
 
-export type TypedArray = Int8Array | Uint8Array | Int16Array | Uint16Array | Int32Array | Uint32Array | Uint8ClampedArray | Float32Array | Float64Array;
-
-export type InputType = string | ArrayBuffer | DataView | TypedArray
-
-export const isTypedArray = (x: any) => (
-  x instanceof Uint8Array ||
-  x instanceof Uint8ClampedArray ||
-  x instanceof Int8Array ||
-  x instanceof Uint16Array ||
-  x instanceof Int16Array ||
-  x instanceof Uint32Array ||
-  x instanceof Int32Array ||
-  x instanceof Float32Array ||
-  x instanceof Float64Array
-);
-
-export enum InputTypes {
-  STRING = 'string',
-  ARRAY_BUFFER = 'arrayBuffer',
-  TYPED_ARRAY = 'typedArray',
-  DATA_VIEW = 'dataView'
-};
-
-const text = {Encoder: TextEncoder, Decoder: TextDecoder};
-
-if (typeof TextEncoder !== 'undefined') {
-  text.Encoder = TextEncoder;
-  text.Decoder = TextDecoder;
-} else {
-  try {
-    // const util = import('util');
-    // text.Encoder = util.TextEncoder;
-    // text.Decoder = util.TextDecoder;
-  } catch (ex) {
-    throw new Error('Arcsecond requires TextEncoder and TextDecoder to be polyfilled.');
-  }
-}
-
-export const encoder = new text.Encoder();
-export const decoder = new text.Decoder();
-
-const getString = (index: number, length: number, dataView: DataView) => {
-  const bytes = Uint8Array.from({length}, (_, i) => dataView.getUint8(index + i));
-  const decodedString = decoder.decode(bytes);
-  return decodedString;
-};
-const getNextCharWidth = (index: number, dataView: DataView) => {
-  const byte = dataView.getUint8(index);
-  if ((byte & 0x80) >> 7 === 0) return 1;
-  else if ((byte & 0xe0) >> 5 === 0b110) return 2;
-  else if ((byte & 0xf0) >> 4 === 0b1110) return 3;
-  else if ((byte & 0xf0) >> 4 === 0b1111) return 4;
-  return 1;
-}
-const getUtf8Char = (index: number, length: number, dataView: DataView) => {
-  const bytes = Uint8Array.from({length}, (_, i) => dataView.getUint8(index + i));
-  return decoder.decode(bytes);
-};
-const getCharacterLength = (str: string) => {
-  let cp;
-  let total = 0;
-  let i = 0;
-  while (i < str.length) {
-    cp = str.codePointAt(i);
-    while (cp) {
-      cp = cp >> 8;
-      i++;
-    }
-    total++;
-  }
-  return total;
-}
-
-//    createParserState :: x -> s -> ParserState e a s
-const createParserState = <D>(target: InputType, data: D | null = null): ParserState<null, string | null, D | null> => {
-  let dataView: DataView;
-  let inputType;
-
-  if (typeof target === 'string') {
-    const bytes = encoder.encode(target);
-    dataView = new DataView(bytes.buffer);
-    inputType = InputTypes.STRING;
-  } else if (target instanceof ArrayBuffer) {
-    dataView = new DataView(target);
-    inputType = InputTypes.ARRAY_BUFFER;
-  } else if (isTypedArray(target)) {
-    dataView = new DataView(target.buffer);
-    inputType = InputTypes.TYPED_ARRAY;
-  } else if (target instanceof DataView) {
-    dataView = target;
-    inputType = InputTypes.DATA_VIEW;
-  } else {
-    throw new Error(`Cannot process input. Must be a string, ArrayBuffer, TypedArray, or DataView. but got ${typeof target}`);
-  }
-
-  return {
-    dataView,
-    inputType,
-
-    isError: false,
-    error: null,
-    result: null,
-    data,
-    index: 0,
-  };
-};
-
-//    updateError :: (ParserState e a s, f) -> ParserState f a s
-const updateError = <T, E, D, E2>(state: ParserState<T, E, D>, error: E2): ParserState<T, E2, D> => ({ ...state, isError: true, error });
-
-//    updateResult :: (ParserState e a s, b) -> ParserState e b s
-const updateResult = <T, E, D, T2>(state: ParserState<T, E, D>, result: T2): ParserState<T2, E, D> => ({ ...state, result });
-
-//    updateData :: (ParserState e a s, t) -> ParserState e b t
-const updateData = <T, E, D, D2>(state: ParserState<T, E, D>, data: D2): ParserState<T, E, D2> => ({ ...state, data });
-
-//    updateResult :: (ParserState e a s, b, Integer) -> ParserState e b s
-const updateParserState = <T, E, D, T2>(state: ParserState<T, E, D>, result: T2, index: number): ParserState<T2, E, D> => ({
-  ...state,
-  result,
-  index,
-});
-
-type StateTransformerFunction<T, E = any, D = null> = (state: ParserState<any, any, any>) => ParserState<T, E, D>;
-export type FnReturingParserIterator<T> = () => Iterator<Parser<any>, T>;
-
-export type ParserState<T, E, D> = {
-  dataView: DataView;
-  inputType: InputType;
-} & InternalResultType<T, E, D>
-
-export type InternalResultType<T, E, D> = {
-  isError: boolean;
-  error: E;
-  index: number;
-  result: T;
-  data: D;
-}
-
-export type ResultType<T, E, D> = Err<E, D> | Ok<T, D>
-
-export type Err<E, D> = {
-  isError: true;
-  error: E;
-  index: number;
-  data: D;
-}
-
-export type Ok <T, D> = {
-  isError: false;
-  index: number;
-  result: T;
-  data: D;
-}
-
-
-
-export class Parser<T, E = string, D = null> {
-    p: StateTransformerFunction<T, E, D>
-
-    constructor(p: StateTransformerFunction<T, E, D>) {
-        this.p = p
-    }
-    
-    //               run :: Parser e a s ~> x -> Either e a
-run(target: InputType): ResultType<T, E, D> {
-    const state = createParserState(target);
-  
-    const resultState = this.p(state);
-  
-    if (resultState.isError) {
-      return {
-        isError: true,
-        error: resultState.error,
-        index: resultState.index,
-        data: resultState.data,
-      }
-    }
-  
-    return {
-      isError: false,
-      result: resultState.result,
-      index: resultState.index,
-      data: resultState.data,
-    };
-  };
-  
-  //               fork :: Parser e a s ~> x -> (e -> ParserState e a s -> f) -> (a -> ParserState e a s -> b)
-  fork<F>(target: InputType, errorFn: (errorMsg: E, parsingState: ParserState<T, E, D>) => F, successFn: (result: T, parsingState: ParserState<T, E, D>) => F) {
-    const state = createParserState(target);
-    const newState = this.p(state);
-  
-    if (newState.isError) {
-      return errorFn(newState.error, newState);
-    }
-  
-    return successFn(newState.result, newState);
-  };
-  
-  //               map :: Parser e a s ~> (a -> b) -> Parser e b s
-  map<T2>(fn: (x: T) => T2): Parser<T | T2, E, D>{
-    const p = this.p;
-    return new Parser(function Parser$map$state(state): ParserState<T | T2, E, D> {
-      const newState = p(state);
-      if (newState.isError) return newState;
-      return updateResult(newState, fn(newState.result));
-    });
-  };
-  
-  //                chain :: Parser e a s ~> (a -> Parser e b s) -> Parser e b s
-  chain<T2>(fn: (x?: T) => Parser<T2, E, D>): Parser<T | T2, E, D> {
-    const p = this.p;
-    return new Parser(function Parser$chain$state(state): ParserState<T | T2, E, D> {
-      const newState = p(state);
-      if (newState.isError) return newState;
-      return fn(newState.result).p(newState);
-    });
-  };
-  
-  //               ap :: Parser e a s ~> Parser e (a -> b) s -> Parser e b s
-  ap<T2>(parserOfFunction: Parser<(x: T) => T2, E, D>): Parser<T2, E, D> {
-    const p = this.p;
-    return new Parser(function Parser$ap$state(state) {
-      if (state.isError) return state;
-  
-      const argumentState = p(state);
-      if (argumentState.isError) return argumentState;
-  
-      const fnState = parserOfFunction.p(argumentState);
-      if (fnState.isError) return fnState;
-  
-      return updateResult(fnState, fnState.result(argumentState.result));
-    });
-  };
-  
-  //               errorMap :: Parser e a s ~> (e -> f) -> Parser f a s
-  errorMap<E2>(fn: (error: Err<E, D>) => E2): Parser<T, E | E2, D> {
-    const p = this.p;
-    return new Parser(function Parser$errorMap$state(state): ParserState<T, E | E2, D> {
-      const nextState = p(state);
-      if (!nextState.isError) return nextState;
-  
-      return updateError(
-        nextState,
-        fn({
-          isError: true,
-          error: nextState.error,
-          index: nextState.index,
-          data: nextState.data
-        }),
-      );
-    });
-  };
-  
-  //               errorChain :: Parser e a s ~> ((e, Integer, s) -> Parser f a s) -> Parser f a s
-  errorChain<T2>(fn: (error: Err<E, D>) => Parser<T2, E, D>): Parser<T | T2, E, D> {
-    const p = this.p;
-    return new Parser(function Parser$errorChain$state(state): ParserState<T | T2, E, D> {
-      const nextState = p(state);
-      if (nextState.isError) {
-        const { error, index, data } = nextState;
-        const nextParser = fn({ isError: true, error, index, data });
-        return nextParser.p({ ...nextState, isError: false });
-      }
-      return nextState;
-    });
-  };
-  
-  //               mapFromData :: Parser e a s ~> (StateData a s -> b) -> Parser e b s
-  mapFromData<T2>(fn: (data: Ok<T, D>) => T2): Parser<T | T2, E, D> {
-    const p = this.p;
-    return new Parser((state): ParserState<T | T2, E, D> => {
-      const newState = p(state);
-      if (newState.isError && newState.error) return newState;
-      return updateResult(
-        newState,
-        fn({ isError: false, result: newState.result, data: newState.data, index: newState.index }),
-      );
-    });
-  };
-  
-  //               chainFromData :: Parser e a s ~> (StateData a s -> Parser f b t) -> Parser f b t
-  chainFromData<T2>(fn: (data: {result: T, data: D}) => Parser<T2, E, D>): Parser<T | T2, E, D> {
-    const p = this.p;
-    return new Parser(function Parser$chainFromData$state(state): ParserState<T | T2, E, D> {
-      const newState = p(state);
-      if (newState.isError && newState.error) return newState;
-      return fn({ result: newState.result, data: newState.data }).p(newState);
-    });
-  };
-  
-  //               mapData :: Parser e a s ~> (s -> t) -> Parser e a t
-  mapData<D2>(fn: (data: D) => D2): Parser<T, E, D2> {
-    const p = this.p;
-    return new Parser(function mapData$state(state) {
-      const newState = p(state);
-      return updateData(newState, fn(newState.data));
-    });
-  };
-  
-  //                   of :: a -> Parser e a s
-  static of<T, E = any, D = null>(x: T): Parser<T, E, D> {
-    return new Parser(state => updateResult(state, x));
-  };
-}
+export {encoder, decoder, Parser, ParserState, ResultType, FnReturingParserIterator, Err, Ok, updateData, updateError, updateParserState, updateResult, InputType, InputTypes, getCharacterLength, getNextCharWidth, getString, getUtf8Char}
 
 //           getData :: Parser e a s
 export function getData() {
@@ -1070,7 +771,7 @@ export const takeLeft = function takeLeft<L, R>(leftParser: Parser<L>) {
 
 //           toPromise :: ParserResult e a s -> Promise (e, Integer, s) a
 export function toPromise<T, E, D>(result: ResultType<T, E, D>) {
-  return result.isError
+  return result.isError === true
     ? Promise.reject({
         error: result.error,
         index: result.index,
@@ -1081,7 +782,7 @@ export function toPromise<T, E, D>(result: ResultType<T, E, D>) {
 
 //           toValue :: ParserResult e a s -> a
 export function toValue<T, E, D>(result: ResultType<T, E, D>): T {
-  if (result.isError) {
+  if (result.isError === true) {
     const e = new Error(String(result.error) || 'null') as any;
     e.parseIndex = result.index; 
     e.data = result.data;
