+++ tsconfig.json
@@ -16,6 +16,6 @@
       }
     },
     "include": [
-      "./index.ts"
+      "src/**/*"
     ]
   }
